fixed one more max_predicted_time overflow
git-svn-id: svn://svn.sphinxsearch.com/sphinx/branches/rel21@3769 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinx.cpp
@@ -17343,7 +17343,7 @@ bool CSphIndex_VLN::ParsedMultiQuery ( const CSphQuery * pQuery, CSphQueryResult
 	// setup prediction constrain
 	CSphQueryStats tQueryStats;
 	bool bCollectPredictionCounters = ( pQuery->m_iMaxPredictedMsec>0 );
-	int64_t iNanoBudget = pQuery->m_iMaxPredictedMsec * 1000000; // from milliseconds to nanoseconds
+	int64_t iNanoBudget = (int64_t)(pQuery->m_iMaxPredictedMsec) * 1000000; // from milliseconds to nanoseconds
 	tQueryStats.m_pNanoBudget = &iNanoBudget;
 	if ( bCollectPredictionCounters )
 		tTermSetup.m_pStats = &tQueryStats;
